Fix type overflow
Summary: The change handles the assert error given by the comparasion between an int and a size_t

Test Plan:
Verified that the change is adequately covered by unit tests.
Performance is not affected.
The solution provided is limited by 2^16 (or greater, depending on the platform in use) number of children.

Reviewers: rhambasan

Reviewed By: rhambasan

Differential Revision: https://test-khuxd457foif.phacility.com/D1